requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
scikit-learn==1.6.1
pandas==2.3.1
joblib
pydantic==1.10.5
httpx
dvc[local]
mlflow==2.8.1
evidently==0.4.11
streamlit-authenticator==0.2.3
plotly==5.17.0
seaborn==0.12.2
numpy==1.23.5